Coalesce cmds, ErrCode, Conn into net
