+++ mpmp/src/net/AddHouse.java
@@ -1,10 +1,8 @@
-package cmds;
+package net;
 
 import java.util.Arrays;
 
 import main.Client;
-import main.Conn;
-import main.ErrCode;
 import model.Player;
 import model.Plot;
 import model.HousePlot;
