+++ mpmp/src/net/StartGame.java
@@ -1,9 +1,10 @@
-package cmds;
+package net;
 
 import java.util.ArrayList;
+import java.util.Random;
+
 import main.Client;
-import main.Conn;
-import main.ErrCode;
+import main.GameState;
 import model.Diceroll;
 import model.Field;
 import model.Player;
@@ -11,24 +12,24 @@ import model.Player;
 /**
  * C->S
  */
-public class EndTurn implements CmdFunc {
+public class StartGame implements CmdFunc {
 
 	@Override
 	public void exec(String line, Conn conn) {
-		// XXX check if plot was bought, if not -> auction
-		Client c = (Client) conn;
-
-		Player p = Player.getCurrentPlayer();
+		GameState.startGame();
+		
 		ArrayList<Player> players = Player.getRealPlayers();
-
-		// Only current player can end their turn.
-		if(!c.getName().equals(p.getName()))
+		Random r = new Random();
+		
+		Player p = Player.search(((Client) conn).getName());
+		if (!p.isPlayer()) {
+			conn.sendErr(ErrCode.NotAPlayer);
 			return;
+		}
 
 		Diceroll dr = new Diceroll();
 		if(dr.getPaschs() >= 3) {
 			p.prison(true);
-			conn.send("prison enter");
 			Client.broadcast("turn-update " + 0 + " " + dr.getPaschs() + " " + p.getName());
 			Client.broadcast("pos-update " + Field.Prison + " " + p.getName());
 		} else {
@@ -38,11 +39,10 @@ public class EndTurn implements CmdFunc {
 		}
 
 		Player.setCurrentPlayer(players.get((players.indexOf(p)+1) % players.size()));
+		conn.sendOK();
 	}
 
 	@Override
-	public void error(ErrCode err, String line, Conn conn) {
-		//TODO
-	}
+	public void error(ErrCode err, String line, Conn conn) {}
 	
 }
