+++ mpmp/src/net/TurnUpdate.java
@@ -1,9 +1,8 @@
-package cmds;
+package net;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import main.Conn;
-import main.ErrCode;
+
 import model.Player;
 import view.Displayer;
